+++ Source/FileBasedApp/FileBasedApp.Toolkit.CSharp/CSharpProjectLoaderExtensions.cs
@@ -5,18 +5,18 @@ namespace FileBasedApp.Toolkit.CSharp;
 
 
 /// <summary>
-/// Provides extension methods for <see cref="CsharpProjectAnalysis"/> to enhance its functionality.
+/// Provides extension methods for <see cref="CSharpProjectAnalysis"/> to enhance its functionality.
 /// </summary>
 /// <remarks>
 /// This static class contains utility extension methods that extend the capabilities of the CsharpProjectLoader class.
 /// </remarks>
-public static class CsharpProjectLoaderExtensions
+public static class CSharpProjectLoaderExtensions
 {
     /// <summary>
-    /// Extension methods for <see cref="CsharpProjectAnalysis"/>
+    /// Extension methods for <see cref="CSharpProjectAnalysis"/>
     /// </summary>
     /// <param name="cls"></param>
-    extension(CsharpProjectAnalysis cls)
+    extension(CSharpProjectAnalysis cls)
     {
         /// <summary>
         /// Finds all named type symbols that implement the specified interface by its fully qualified metadata name.